feat:refactored addresses
