+++ src/context/address/Context.jsx
@@ -1,6 +1,6 @@
 import { useContext, useEffect, useReducer, createContext } from "react";
 import { addressReducer } from "./Reducer";
-import { getUserAddressHandler, addNewAddressHandler } from "./services/utils";
+import { getUserAddressHandler, addNewAddressHandler } from "./utils/services";
 
 const AddressContext = createContext();
