+++ src/context/address/Reducer.jsx
@@ -1,7 +1,9 @@
+import { GET_ADDRESS, ADD_ADDRESS } from "./utils/constants";
+
 const addressReducer = (state, action) => {
   switch (action.type) {
-    case "GET_ADDRESS":
-    case "ADD_ADDRESS":
+    case GET_ADDRESS:
+    case ADD_ADDRESS:
       return {
         ...state,
         address: action.payload,
