+++ src/context/address/utils/constants.jsx
@@ -1,0 +1,2 @@
+export const GET_ADDRESS = "GET_ADDRESS";
+export const ADD_ADDRESS = "ADD_ADDRESS";
